cmd,Makefile: Introduce wayfinder Daemon program
This commit re-organises the build system to introduce a new binary
output, `wayfinderd`, which will perform the main actions.  The
daemon relies on [0] to faciliate multi-async modules to run and
receive requests.  It provides gRPC, HTTP and additional modules for
quick access to other distributed systems, e.g. postgres, redis.

Signed-off-by: Alexander Jung <alex@unikraft.io>