spec: Move specification to its own module
By moving the YAML declaration of a Wayfinder job to its own module,
it abstracts the definition and allows it to be used in other places.
Additionally, operation-specific commands on the file can be kept
within the module, for example, creating permutations based on the
parameter definitions.

This commit moves the specification to its own module and adjusts
the specification at the same time, creating a clear saparation
between a "build" and a "test".  The build now has now two specific
outputs: the kernel binary and the initramfs.  The test has been
well-defined also, and includes information about how to instantiate
the kernel, i.e. with kernel-line arguments, as well as how to test
the kernel with a pre-define OCI image.  The OCI image is a
"benchmarking tool" (benchtool) and now has an added start-delay to
allow for the kernel to warm up.

This new specification relies on the protocols defined in api/

Signed-off-by: Alexander Jung <alex@unikraft.io>