config: Reduce pre-fetch limit to 1
The prefetch limit should be set to 1 so that the jobs and permutation
task list can be correctly resumed if unexpectedly quit.  This means that
the queue will hold in redis and when the daemon restarts, it will
resume correctly.  For now, there is only 1 consumer instantiated per
job/permutation.  In the future, it may be possible to set this number
higher and then imply more concurrent handlers which are able to handle
allocation of cores correctly.

Signed-off-by: Alexander Jung <alex@unikraft.io>