internal/coremap: Fix searching for cores by ID
This commit fixes the generic function to serach for a Core by its
ID.  The function returned early during the loop through all of the
NUMA nodes, meaning it would only search the first NUMA node.  In
many cases, this would mean the core was not found if there is more
than one node.

Signed-off-by: Alexander Jung <alex@unikraft.io>