modules/scheduler: Abstract the logic to reserve cores
This commit introduces a new method which is used to reserve CPU cores
which is used by both 'build' and 'test' stage of a job's scheduling.
The method is abstract in that it takes as input the number of cores
desired as well as a pointer to an interface, and then busy-waits to
reserve the number of cores required, setting them as busy and
returning the core IDs.  This is then used by the respective part of
the scheduler.

Signed-off-by: Alexander Jung <alex@unikraft.io>