*: Move evaluation of expression to job generation
The old position of the evaluation would only empty the values.
This means that there would be duplicate permutations (the ones that
have been emptied) when running a big job.

Moving the evaluation right before adding it to the redis queue enables
us to skip saving permutations that are not unique (that do not have an
unique checksum).

As a final mention, the condition evaluation does not depend on the
order of parameters.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>