internal/coremap: Refactor coremap to use layout
By using the command 'lscpu --all --parse' we can extract information
about the CPU much easier. With this we can further expand the
abstractisation from Numa/CPU to Socket/Numa/Cache/CPU.

With this new format we can search for more specific cores when looking
for new available free cores.

Note that the cache groups and the numa nodes contain the same cores,
so, when reserving a core from one of them, it is marked as busy in the
other too.

All the changes made should retain the same functionality of wayfinder.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>