api,modules/container,spec: Introduce ability to specify workdir
This commit introduces the ability to specify a working directory
within the container environment created by wayfinder.  Relying on
the `ImageConfig` method, we are able to either automatically set
the working directory set within the OCI config or override this
path to set the `cwd` of the set of commands run within the
environment.

Signed-off-by: Alexander Jung <alex@unikraft.io>