modules/builder,modules/container: Always attach devices and return err
This commit forces devices to always be attached.  The previous
functionality would mean that no devices would be attached if either
the `build` or `test` directive of a configuration file ommited
the `devices` attribute.  For the container runtime environment
to be created, a base set of devices must be attached always.  This
commit fixes this by calling the `SetDevices` method always and
allows for returning an error if it occurs from this operation.

Signed-off-by: Alexander Jung <alex@unikraft.io>