modules/container: Introduce special access to loop devices
This commit introduces the ability to attach loop devices to the
container environment.  When the special device `/dev/loop-control`
is detected as part of the listed devices for a container environment
(either in the `test` or `build` of a configuration), we automatically
enumarate available loop devices (`/dev/loopN`) on the host and
attach them appropriately.  This is useful for running commands like
`mount -o loop` within the context of a build.

Signed-off-by: Alexander Jung <alex@unikraft.io>