internal/repositories: Fix preloading of related table data
This commit fixes the preloading of the associated tables data.  For
now, we preload all the data so that the API can expose all the data
to be used for analysis.  Preloading is however not efficient as it
creates very large queries with over 200ms of execution time and
very large network data transmissions.  In the future, lazy-loading
of data should be possible, especially via the client libraries
which have contextual information about the data saved to them.

Signed-off-by: Alexander Jung <alex@unikraft.io>